refactor: rename config files and remove unused imports
